+++ src/DiplomaticAction/NonAggressionPact/NonAggressionPactScoringModel.cs
@@ -14,7 +14,7 @@ namespace Diplomacy.DiplomaticAction.NonAggressionPact
             public int ExistingAllianceWithEnemy => -1000;
             public int ExistingAllianceWithNeutral => -50;
             public int Relationship => 50;
-            public float Tendency => Settings.Instance!.NonAggressionPactTendency;
+            public int Tendency => Settings.Instance!.NonAggressionPactTendency;
             public int NonAggressionPactWithEnemy => -20;
             public int NonAggressionPactWithNeutral => -10;
         }
